# server/openapi.yaml
openapi: "3.0.3"
info:
  title: ImmoChecker API
  version: "1.0.0"
  description: |
    API publique ImmoChecker pour créer et consulter des missions, EDL, lots et bâtiments.

    ## Authentification
    Toutes les requêtes nécessitent une clé API créée depuis **Paramètres → API & Intégrations**.

    ```
    Authorization: Bearer imk_live_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
    ```

    ## Permissions (scopes)
    Chaque clé API a un scope :
    - **`read`** — lecture seule (tous les `GET`)
    - **`read_write`** — lecture + écriture (`POST`, `PATCH`, `DELETE`)

    Une écriture avec une clé `read` renvoie un `403 FORBIDDEN_SCOPE`.

    ## Rate Limiting
    100 requêtes par minute par clé API. Les headers `RateLimit-*` indiquent les limites restantes.
    Dépassement → `429 RATE_LIMIT`.

    ## Pagination
    Toutes les listes sont paginées avec un curseur :
    ```
    GET /missions?limit=25&cursor=<id>
    ```
    La réponse contient `meta.cursor` (prochain curseur, `null` si fin de liste) et `meta.has_more`.
    `limit` par défaut = 25, max = 100.

    ## Webhooks
    Les webhooks (`mission.creee`, `mission.annulee`, `edl.signe`, …) ne se configurent pas via cette
    API publique mais depuis **Paramètres → API & Intégrations** (session admin). Pour la documentation
    du payload et de la signature HMAC, consulter le panneau Webhooks dans le back-office.

    ## ⚠️ Try it out
...
servers:
  - url: https://immo-check-production.up.railway.app/api/v1
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "imk_live_..."
      description: Clé API créée depuis Paramètres → API & Intégrations
  schemas:
    Meta:
      type: object
      properties:
        cursor:
          type: string
          nullable: true
          description: Curseur pour la page suivante (`null` si fin de liste)
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        has_more:
          type: boolean
          example: true
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Message humain en français
          example: "Ressource introuvable"
        code:
          type: string
          description: Code machine stable (à utiliser en logique de branchement)
          example: "NOT_FOUND"
        details:
          description: Détails additionnels (ex. erreurs de validation)
          nullable: true
    Lot:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        designation:
          type: string
          example: "Appartement T3 — Lot 302"
        type_bien:
          type: string
          enum: [appartement, maison, local_commercial, bureau, parking, cave, autre]
          example: "appartement"
        etage:
          type: string
          nullable: true
          example: "3ème"
        surface:
          type: number
          nullable: true
          example: 67
        reference_interne:
          type: string
          nullable: true
          example: "BAIL-2025-0302"
        batiment:
          type: object
          properties:
            id:
              type: string
              format: uuid
              example: "b23f1a2c-9d8e-4b3a-a1c2-d3e4f5a6b7c8"
            designation:
              type: string
              example: "Résidence Victor Hugo"
            adresse:
              type: string
              example: "12 rue Victor Hugo, 75016 Paris"
...
    PaginatedLots:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Lot' }
        meta: { $ref: '#/components/schemas/Meta' }
...
  responses:
    Unauthorized:
      description: Clé API manquante ou invalide
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            missing:
              summary: En-tête Authorization absent
              value: { error: "Clé API requise", code: "UNAUTHORIZED" }
            invalid:
              summary: Clé inconnue ou révoquée
              value: { error: "Clé API invalide", code: "INVALID_API_KEY" }
...
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            default:
              value: { error: "Mission introuvable", code: "NOT_FOUND" }
...
paths:
...
  /lots:
    get:
      summary: Lister les lots
      operationId: listLots
      tags: [Référentiel]
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: batiment_id
          schema: { type: string, format: uuid }
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 25, maximum: 100 }
      responses:
        '200':
          description: Liste paginée de lots
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedLots' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /lots/{id}:
    get:
      summary: Détail d'un lot
      operationId: getLot
      tags: [Référentiel]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Lot
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lot' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...